Change ClickHouse.Client to ClickHouse.Driver (#5071)
* Change ClickHouse.Client to ClickHouse.Driver and attendant assembly names

ClickHouse is officially supporting a fork of the original ClickHouse.Client, changing all namespaces including the assembly and package names to ClickHouse.Driver.

This breaks the linq2db integration on the ClickHouse side, as the library tries to resolve ClickHouse.Client.dll and other namespaces that have changed to use Driver nomenclature.

The commit should switch these references to the officially supported client.

* Fix ProviderName for AllClickHouse

* Fix additional name references from ClickHouseClient to ClickHouseDriver

* - fix ci

- remove old versions compatibility code
- additional renames
- revert some wrong renames

* fix adapter wrappers registration

---------

Co-authored-by: MaceWindu <MaceWindu@users.noreply.github.com>